Change randomness behaviour to be like it prior to go1.20
Pre go1.20, rand.Seed needed to be called for randomness. Meaning, that
math/rand delivered deterministic values by default. This was changed
however. To restore the old behaviour, we need to manually seed our
random generator with a constant.

Additionally, this introduce a local instance of rand.Rand, in order
to prevent unexpected behaviour with other code that relies on the
global math/rand package.

Fixes #10 (The readme and docs explain this, therefore I consider this
solved, as it is by design and not a bug)

Fixes #15